Create BioFace: Age, Gender, BMI Predictor
The project uses facial images as input to predict an individual's age, gender, BMI, height, and weight. It utilizes the FaceRecognition library for face detection, alignment, and feature extraction from the input images. Subsequently, a Kernel Ridge Regression (KRR) model is employed to analyze the extracted facial features and make predictions of the subject's age, gender, weight, height, and BMI.